refactor: 이미지 저장 함수 분리
